requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
et_xmlfile==2.0.0
google-analytics-data==0.18.18
google-api-core==2.25.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
grpcio==1.72.1
grpcio-status==1.72.1
idna==3.10
mysqlclient==2.2.7
numpy==2.2.6
oauthlib==3.2.2
openpyxl==3.1.5
pandas==2.3.0
pillow==11.2.1
proto-plus==1.26.1
protobuf==6.31.1
py4j==0.10.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.9.0
pyspark==4.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
reportlab==4.4.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
